Improve handling of incomplete senders
The official way to represent an IRC user is `nickname!ident@host`
However, some IRC servers will omit the `!ident@host` part for various
reasons, sending only a nickname. Previously, IRCSharp would consider
that to be an IRC user with no nickname, no ident, and only a host
portion, because the 'nickname' sent by the server would usually look
like `irc.server.tld`.

But, as it turns out, some servers will represent service users in
the same way, sending `service` instead of `service!ident@host`.
In this case it obviously makes no sense to consider `service` to be
the hostname, and, in retrospect, it also makes more sense to consider
`irc.server.tld` to be a nickname, rather than a host.

So when a server sends only `nickname`, `nickname` will now be
considered to be that user's nickname, rather than their host.